application/pyproject.toml
[project]
name = "vapar"
version = "0.1.0"
description = ""
authors = [
    {name = "Sam Dyson",email = "mail@samdyson.com"}
]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13.6"

django = ">=5.2.5,<6.0.0"
djangorestframework = ">=3.16.1,<4.0.0"
adrf = "^0.1.9"
httpx = ">=0.28.1,<0.29.0"
daphne = "^4.2.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "vapar.settings.local"
pythonpath = "vapar tests"
python_files = "tests/**/test_*.py"
addopts = "-p no:warnings --cov=. "
